+++ rpc/core/types/responses.go
@@ -1,8 +1,8 @@
 package core_types
 
 import (
-	acm "github.com/eris-ltd/eris-db/account"
-	stypes "github.com/eris-ltd/eris-db/state/types"
+	acm "github.com/eris-ltd/eris-db/manager/eris-mint/account"
+	stypes "github.com/eris-ltd/eris-db/manager/eris-mint/state/types"
 	txtypes "github.com/eris-ltd/eris-db/txs"
 	"github.com/tendermint/tendermint/types"
